core: fix format specifier related errors

+++ core/src/lib/edit.cc
@@ -279,7 +279,7 @@ static std::pair<std::uint64_t, const char*> parse_number_with_mod(
         }
       }
       if (!found) {
-        Dmsg1(900, "Unknown modifier: \"%.*s\"\n", modifier.size(),
+        Dmsg1(900, "Unknown modifier: \"%.*s\"\n", (int)modifier.size(),
               modifier.data());
         return {total, str};
       }
